requirements.txt
pre-commit
python-dotenv
requests
